Updating to latest FlutterFlow output.

+++ lib/profile/email_template_editor/email_template_editor_widget.dart
@@ -153,8 +153,33 @@ class _EmailTemplateEditorWidgetState extends State<EmailTemplateEditorWidget> {
                     child: FFButtonWidget(
                       onPressed: () async {
                         if (widget.emailTemp != null) {
-                          await widget.emailTemp!.reference.delete();
-                          context.safePop();
+                          var confirmDialogResponse = await showDialog<bool>(
+                                context: context,
+                                builder: (alertDialogContext) {
+                                  return AlertDialog(
+                                    title: const Text('Delete Template'),
+                                    content: const Text(
+                                        'Are you sure you want to delete the template'),
+                                    actions: [
+                                      TextButton(
+                                        onPressed: () => Navigator.pop(
+                                            alertDialogContext, false),
+                                        child: const Text('Cancel'),
+                                      ),
+                                      TextButton(
+                                        onPressed: () => Navigator.pop(
+                                            alertDialogContext, true),
+                                        child: const Text('Confirm'),
+                                      ),
+                                    ],
+                                  );
+                                },
+                              ) ??
+                              false;
+                          if (confirmDialogResponse) {
+                            await widget.emailTemp!.reference.delete();
+                            context.safePop();
+                          }
                         }
                       },
                       text: 'Delete Template',
